Change kertaaviikossa type to smallint

+++ db/migrations/V2.26.0__Update_tyhjennysvali.sql
@@ -1,2 +1,2 @@
 ALTER TABLE jkr.tyhjennysvali
-    ADD COLUMN kertaaviikossa INTEGER;
+    ADD COLUMN kertaaviikossa smallint;
